chore: refactor transaction seeder
