+++ test/unit/service/debtor-service.ts
@@ -18,12 +18,8 @@
 
 import { Connection } from 'typeorm';
 import User, { TermsOfServiceStatus, UserType } from '../../../src/entity/user/user';
-import ProductRevision from '../../../src/entity/product/product-revision';
-import ContainerRevision from '../../../src/entity/container/container-revision';
-import PointOfSaleRevision from '../../../src/entity/point-of-sale/point-of-sale-revision';
 import {
   seedFines,
-  seedTransactions,
   seedTransfers,
 } from '../../seed-legacy';
 import Database from '../../../src/database/database';
@@ -46,15 +42,12 @@ import { finishTestDB } from '../../helpers/test-helpers';
 import dinero from 'dinero.js';
 import TransferService from '../../../src/service/transfer-service';
 import FineHandoutEvent from '../../../src/entity/fine/fineHandoutEvent';
-import { ContainerSeeder, PointOfSaleSeeder, ProductSeeder, UserSeeder } from '../../seed';
+import { TransactionSeeder, UserSeeder } from '../../seed';
 
 describe('DebtorService', (): void => {
   let ctx: {
     connection: Connection,
     users: User[],
-    productRevisions: ProductRevision[],
-    containerRevisions: ContainerRevision[],
-    pointOfSaleRevisions: PointOfSaleRevision[],
     transactions: Transaction[],
     subTransactions: SubTransaction[],
     transfers: Transfer[],
@@ -72,10 +65,7 @@ describe('DebtorService', (): void => {
     await truncateAllTables(connection);
 
     const users = await new UserSeeder().seedUsers();
-    const { productRevisions } = await new ProductSeeder().seedProducts(users);
-    const { containerRevisions } = await new ContainerSeeder().seedContainers(users, productRevisions);
-    const { pointOfSaleRevisions } = await new PointOfSaleSeeder().seedPointsOfSale(users, containerRevisions);
-    const { transactions } = await seedTransactions(users, pointOfSaleRevisions, new Date('2020-02-12'), new Date('2021-11-30'), 10);
+    const { transactions } = await new TransactionSeeder().seedTransactions(users, undefined, new Date('2020-02-12'), new Date('2021-11-30'), 10);
     const transfers = await seedTransfers(users, new Date('2020-02-12'), new Date('2021-11-30'));
     const subTransactions: SubTransaction[] = Array.prototype.concat(...transactions
       .map((t) => t.subTransactions));
@@ -84,9 +74,6 @@ describe('DebtorService', (): void => {
     ctx = {
       connection,
       users: usersWithFines,
-      productRevisions,
-      containerRevisions,
-      pointOfSaleRevisions,
       transactions,
       subTransactions,
       transfers,
