+++ test/unit/service/vat-group-service.ts
@@ -23,14 +23,18 @@ import User from '../../../src/entity/user/user';
 import Transaction from '../../../src/entity/transactions/transaction';
 import { VatGroupRequest } from '../../../src/controller/request/vat-group-request';
 import Database from '../../../src/database/database';
-import {
-  seedTransactions,
-} from '../../seed-legacy';
 import VatGroupService from '../../../src/service/vat-group-service';
 import { VatDeclarationResponse } from '../../../src/controller/response/vat-group-response';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
-import { ContainerSeeder, PointOfSaleSeeder, ProductSeeder, UserSeeder, VatGroupSeeder } from '../../seed';
+import {
+  ContainerSeeder,
+  PointOfSaleSeeder,
+  ProductSeeder,
+  TransactionSeeder,
+  UserSeeder,
+  VatGroupSeeder,
+} from '../../seed';
 
 describe('VatGroupService', () => {
   let ctx: {
@@ -49,7 +53,7 @@ describe('VatGroupService', () => {
     const { productRevisions } = await new ProductSeeder().seedProducts(users, undefined, vatGroups, 100);
     const { containerRevisions } = await new ContainerSeeder().seedContainers(users, productRevisions);
     const { pointOfSaleRevisions } = await new PointOfSaleSeeder().seedPointsOfSale(users, containerRevisions);
-    const { transactions } = await seedTransactions(users, pointOfSaleRevisions, new Date('2020-02-12'), new Date('2022-11-30'), 3);
+    const { transactions } = await new TransactionSeeder().seedTransactions(users, pointOfSaleRevisions, new Date('2020-02-12'), new Date('2022-11-30'), 3);
 
     const validVatCreateReq: VatGroupRequest = {
       name: 'Extreem hoog tarief',
